extract out dice generators
